Fix dropout_buffer allocated with sizeof(float) instead of sizeof(uint8_t)
The detect_dropouts shader writes uint8_t values, so the intermediate
dropout_buffer in both demodulators should be byte-sized. Using sizeof(float)
made it 4x larger than necessary without affecting correctness.

+++ player/src/muse/MuseRfDemodulator.cpp
@@ -106,7 +106,7 @@ void MuseRfDemodulator::demodulate() {
             m_vulkan_manager, Size(rrc_in_buffer_size), sizeof(float), buffer_usage_flags, HostAccess::eHostNone);
 
     shared_ptr<VulkanBuffer> dropout_buffer = make_unique<musevk::VulkanBuffer>(
-            m_vulkan_manager, Size(c_dropout_buffer_size), sizeof(float), buffer_usage_flags, HostAccess::eHostNone);
+            m_vulkan_manager, Size(c_dropout_buffer_size), sizeof(uint8_t), buffer_usage_flags, HostAccess::eHostNone);
 
     // Create shaders
     shared_ptr<ComputeShader> input_fir_filter_shader = unique_ptr<ComputeShader>(
